Fix/clean is_loop_carried & add is_loop_independent function
The code in is_loop_carried checking if the loop is independent
has been removed (because not necessary).
An osl_scop is no longer necessary for the function btw (all call to
this function has been modified to remove the osl_scop param).

A function candl_dependence_is_loop_independent has beed added,
that add some constraint and use piplib to check if the dependence
is loop independent or not.